add flags to compile as c++ code
tried the plf_nanotimer library which is in c++, so temporarily compiled the .c files as c++ source

+++ src/vector.c
@@ -5,7 +5,7 @@
 // Implementations of Vector 2D functions
 ///////////////////////////////////////////////////////////////////////////////
 float vec2_length(vec2_t v) {
-    return sqrt(v.x * v.x + v.y * v.y);
+    return sqrtf(v.x * v.x + v.y * v.y);
 }
 
 vec2_t vec2_add(vec2_t a, vec2_t b) {
@@ -48,7 +48,7 @@ float vec2_dot(vec2_t a, vec2_t b) {
 // Implementations of Vector 3D functions
 ///////////////////////////////////////////////////////////////////////////////
 float vec3_length(vec3_t v) {
-    return sqrt(v.x * v.x + v.y * v.y + v.z * v.z);
+    return sqrtf(v.x * v.x + v.y * v.y + v.z * v.z);
 }
 
 vec3_t vec3_add(vec3_t a, vec3_t b) {
@@ -105,7 +105,7 @@ void vec2_normalize(vec2_t* a)
     float lsq = vec2_dot(*a, *a);
     if(lsq)
     {
-        float rlen = 1.0f / sqrt(lsq);
+        float rlen = 1.0f / sqrtf(lsq);
         a->x *= rlen;
         a->y *= rlen;
     }
@@ -117,7 +117,7 @@ void vec3_normalize(vec3_t* a)
     float lsq = vec3_dot(*a, *a);
     if(lsq)
     {
-        float rlen = 1.0f / sqrt(lsq);
+        float rlen = 1.0f / sqrtf(lsq);
         a->x *= rlen;
         a->y *= rlen;
         a->z *= rlen;
@@ -127,25 +127,25 @@ void vec3_normalize(vec3_t* a)
 vec3_t vec3_rotate_x(vec3_t v, float angle) {
     vec3_t rotated_vector = {
         .x = v.x,
-        .y = v.y * cos(angle) - v.z * sin(angle),
-        .z = v.y * sin(angle) + v.z * cos(angle)
+        .y = v.y * cosf(angle) - v.z * sinf(angle),
+        .z = v.y * sinf(angle) + v.z * cosf(angle)
     };
     return rotated_vector;
 }
 
 vec3_t vec3_rotate_y(vec3_t v, float angle) {
     vec3_t rotated_vector = {
-        .x = v.x * cos(angle) - v.z * sin(angle),
+        .x = v.x * cosf(angle) - v.z * sinf(angle),
         .y = v.y,
-        .z = v.x * sin(angle) + v.z * cos(angle)
+        .z = v.x * sinf(angle) + v.z * cosf(angle)
     };
     return rotated_vector;
 }
 
 vec3_t vec3_rotate_z(vec3_t v, float angle) {
     vec3_t rotated_vector = {
-        .x = v.x * cos(angle) - v.y * sin(angle),
-        .y = v.x * sin(angle) + v.y * cos(angle),
+        .x = v.x * cosf(angle) - v.y * sinf(angle),
+        .y = v.x * sinf(angle) + v.y * cosf(angle),
         .z = v.z
     };
     return rotated_vector;
